requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
joblib==1.4.2
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
